requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
chardet==5.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==5.1.6
django-allauth==65.4.1
django-tailwind==3.8.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
fonttools==4.56.0
kiwisolver==1.4.8
kombu==5.4.2
matplotlib==3.10.0
mysqlclient==2.2.7
narwhals==1.26.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
prompt_toolkit==3.0.50
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
reportlab==4.3.1
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
